src/quantum_consciousness/core/quantum_hybrid_cognitive.py
"""
Quantum hybrid cognitive architecture implementation.
Integrates quantum and classical processing for cognitive modeling.
"""

import numpy as np
from typing import List, Tuple, Dict, Optional, Callable
from scipy.linalg import expm, sqrtm
from dataclasses import dataclass

@dataclass
class CognitiveState:
    """Represents a hybrid quantum-classical cognitive state."""
    quantum_state: np.ndarray
    classical_state: np.ndarray
    attention_mask: np.ndarray

class QuantumHybridCognitive:
    """Implements hybrid quantum-classical cognitive architecture."""

    def __init__(self, quantum_dim: int, classical_dim: int):
        """
        Initialize hybrid cognitive system.

        Args:
            quantum_dim: Dimension of quantum subsystem
            classical_dim: Dimension of classical subsystem
        """
        self.quantum_dim = quantum_dim
        self.classical_dim = classical_dim
        self.epsilon = 1e-12

    def initialize_cognitive_state(self) -> CognitiveState:
        """
        Initialize hybrid cognitive state
        |ψ⟩ ⊗ |c⟩

        Returns:
            CognitiveState: Initial hybrid state
        """
        # Initialize quantum state in superposition
        quantum_state = np.ones(self.quantum_dim) / np.sqrt(self.quantum_dim)

        # Initialize classical state
        classical_state = np.zeros(self.classical_dim)
        classical_state[0] = 1

        # Initialize attention mask
        attention_mask = np.ones(self.quantum_dim)

        return CognitiveState(
            quantum_state=quantum_state,
            classical_state=classical_state,
            attention_mask=attention_mask
        )

    def quantum_perception(self, state: CognitiveState,
                         stimulus: np.ndarray) -> CognitiveState:
        """
        Process quantum perception
        P_q(|ψ⟩) = U_p|ψ⟩

        Args:
            state: Current cognitive state
            stimulus: Input stimulus

        Returns:
            CognitiveState: Updated state after perception
        """
        # Create perception operator
        U_p = self._create_perception_operator(stimulus)

        # Apply quantum transformation
        new_quantum_state = U_p @ state.quantum_state

        # Update attention based on stimulus
        new_attention = self._update_attention(state.attention_mask, stimulus)

        return CognitiveState(
            quantum_state=new_quantum_state,
            classical_state=state.classical_state,
            attention_mask=new_attention
        )

    def _create_perception_operator(self, stimulus: np.ndarray) -> np.ndarray:
        """Create unitary perception operator."""
        # Generate Hermitian matrix from stimulus
        H = stimulus @ stimulus.conj().T
        # Create unitary operator
        return expm(-1j * H)

    def _update_attention(self, current_attention: np.ndarray,
                        stimulus: np.ndarray) -> np.ndarray:
        """Update attention mask based on stimulus."""
        # Calculate attention weights
        weights = np.abs(stimulus)
        # Normalize and combine with current attention
        new_attention = 0.7 * current_attention + 0.3 * weights
        return new_attention / np.linalg.norm(new_attention)

    def classical_processing(self, state: CognitiveState) -> CognitiveState:
        """
        Apply classical processing
        C(x) = σ(Wx + b)

        Args:
            state: Current cognitive state

        Returns:
            CognitiveState: Updated state after classical processing
        """
        # Create classical neural network layer
        W = np.random.randn(self.classical_dim, self.classical_dim)
        b = np.random.randn(self.classical_dim)

        # Apply classical transformation
        new_classical_state = np.tanh(W @ state.classical_state + b)

        return CognitiveState(
            quantum_state=state.quantum_state,
            classical_state=new_classical_state,
            attention_mask=state.attention_mask
        )

    def quantum_cup_product(self, state: CognitiveState,
                          memory: np.ndarray) -> np.ndarray:
        """
        Compute quantum cup product for cognitive integration
        a ∗_q b ≈ (Perception ∪ Attention)β q^{⟨β,ω⟩} (Memory)

        Args:
            state: Current cognitive state
            memory: Quantum memory state

        Returns:
            numpy.ndarray: Integrated cognitive state
        """
        # Calculate quantum perception-attention coupling
        perception = state.quantum_state
        attention = state.attention_mask
        beta = 1.0  # Coupling strength

        # Compute overlap
        omega = np.vdot(perception, memory)

        # Calculate cup product
        combined_state = np.kron(perception, attention)
        coupling_factor = np.exp(beta * omega)

        return coupling_factor * (combined_state @ memory)

    def measure_cognitive_coherence(self, state: CognitiveState) -> float:
        """
        Measure cognitive coherence using quantum Fisher information
        F_Q(ρ, A) = 2∑_{i,j} |⟨i|A|j⟩|²/(λᵢ+λⱼ)

        Args:
            state: Cognitive state

        Returns:
            float: Cognitive coherence measure
        """
        # Create density matrix
        rho = np.outer(state.quantum_state, state.quantum_state.conj())


        # Create observable (attention-weighted)
        A = np.diag(state.attention_mask)

        # Calculate eigendecomposition
        eigenvals, eigenvecs = np.linalg.eigh(rho)
        eigenvals = eigenvals[eigenvals > self.epsilon]

        # Calculate quantum Fisher information
        F_Q = 0
        for i, λi in enumerate(eigenvals):
            for j, λj in enumerate(eigenvals):
                if λi + λj > self.epsilon:
                    F_Q += np.abs(np.vdot(eigenvecs[:,i], A @ eigenvecs[:,j]))**2 / (λi + λj)

        return 2 * F_Q

    def integrate_quantum_classical(self, state: CognitiveState) -> CognitiveState:
        """
        Integrate quantum and classical information
        |ψ'⟩ = U_int(|ψ⟩ ⊗ |c⟩)

        Args:
            state: Current cognitive state

        Returns:
            CognitiveState: Integrated state
        """
        # Create interaction unitary
        U_int = self._create_interaction_unitary()

        # Create joint state
        joint_state = np.kron(state.quantum_state, state.classical_state)

        # Apply interaction
        evolved_state = U_int @ joint_state

        # Extract new quantum and classical states
        new_quantum_state = evolved_state[:self.quantum_dim]
        new_classical_state = evolved_state[self.quantum_dim:]

        return CognitiveState(
            quantum_state=new_quantum_state / np.linalg.norm(new_quantum_state),
            classical_state=new_classical_state / np.linalg.norm(new_classical_state),
            attention_mask=state.attention_mask
        )

    def _create_interaction_unitary(self) -> np.ndarray:
        """Create unitary operator for quantum-classical interaction."""
        # Generate random Hermitian matrix
        H = np.random.randn(self.quantum_dim + self.classical_dim,
                          self.quantum_dim + self.classical_dim)
        H = H + H.conj().T
        # Create unitary operator
        return expm(-1j * H)

    def cognitive_feedback_loop(self, state: CognitiveState,
                              iterations: int) -> List[float]:
        """
        Implement cognitive feedback loop with coherence monitoring
        |ψ(t+1)⟩ = F(|ψ(t)⟩)

        Args:
            state: Initial cognitive state
            iterations: Number of feedback iterations

        Returns:
            List[float]: Coherence measures over iterations
        """
        coherence_history = []
        current_state = state

        for _ in range(iterations):
            # Measure current coherence
            coherence = self.measure_cognitive_coherence(current_state)
            coherence_history.append(coherence)

            # Apply cognitive transformations
            current_state = self.quantum_perception(
                current_state,
                np.random.randn(self.quantum_dim)
            )
            current_state = self.classical_processing(current_state)
            current_state = self.integrate_quantum_classical(current_state)

        return coherence_history
